refactor: inherits parent class in data loader and preprocessor
